+++ drift_dev/test/analysis/resolver/drift/table_test.dart
@@ -295,4 +295,21 @@ class MyType implements CustomSqlType<String> {}
         expect(custom.expression.toString(), 'MyType()');
     }
   });
+
+  test('recognizes bigint columns', () async {
+    final state = await TestBackend.inTest({
+      'a|lib/a.drift': '''
+CREATE TABLE foo (
+  bar INT64 NOT NULL
+);
+''',
+    });
+
+    final file = await state.analyze('package:a/a.drift');
+    state.expectNoErrors();
+
+    final table = file.analyzedElements.single as DriftTable;
+    final column = table.columns.single;
+    expect(column.sqlType.builtin, DriftSqlType.bigInt);
+  });
 }